Fit Combined Models with TMB

Fit Hierarchical models with normal and conjugate random effects with the TMB (Template Model Builder) R package. This can be to model hierarchical data subject to within-unit correlation and/or overdispersion.

Usage

combTMB(
  formula,
  data = NULL,
  family = gaussian(link = "identity"),
  dformula = ~1,
  REML = FALSE,
  doMarginal = FALSE,
  weights = NULL,
  offset = NULL,
  contrasts = NULL,
  na.action,
  starting_val = "zero",
  control = combTMBcontrol(),
  map = NULL,
  size = NULL,
  dofit = TRUE,
  silent = TRUE
)

Arguments

formula

combined fixed and random effects formula, following lme4 syntax. IID random intercepts are possible using e.g., `+ (1 | Subject)` where `Subject` is a factor representing groups.

data

an optional data frame containing the variables in the model. If missing data, the variables ara taken from environment(formula). Not required, but strongly recommended. If not specified, a warning message will be issued.

family

a family function, a character string naming a family function. Options are [gaussian()], [binomial()], [poisson()], Beta(), betabinomial(), poigamma(), bb() and bb_al(). For binomial family options, see 'Binomial families' in the Details section below. For more details see family_combTMB.

dformula

an object of class "formula" describe the model for dispersion containing only fixed effects: the default ~1: specifies the standard dispersion any family, however the argument is ignored for families that do not have a dispersion parameter, for example [binomial()] family. One can declare a linear predictor for the dispersion, internally a log link is applied.

REML

logical: use REML (restricted maximum likelihood) estimation rather than maximum likelihood. Internally, this adds the fixed effects to the list of random effects to integrate over.

doMarginal

marginalization of regression parameters is only available for a random effect for the following families: [binomial()], [poisson()] and poigamma(). The default NULL. The argument is ignored for the other families.

weights

an optional vector representing likelihood weights for the conditional model. Weights are not modified internally and do not scale to sum 1. The default is NULL. Can also be used for testing with the binomial family - see Details. The `weights` argument needs to be a vector and not a name of the variable in the data frame

offset

a numeric vector representing the model offset. The `offset` argument needs to be a vector and not a name of the variable in the data frame. *Not included in any prediction.*

contrasts

an optional list. See the contrasts.arg of model.matrix.default.

na.action

a function that specifies how to handle observations containing NAs. The default action (na.omit, inherited from the 'factory fresh' value of getOption("na.action")) strips any observations with any missing values in any variables. Using na.action = na.exclude.

starting_val

an character indicating which method should be used to provide initial values. Options are the estimated values through a fitted model using glm.fit "fm" and all initial parameter values are zero "zero".

control

optimization control options via combTMBcontrol.

map

a named list where the user can set a parameter to a constant value instead of estimating it. A fixed value is the one established as the initial guess in the optimization process. To understand the use of the map argument consider the example, map=list(beta=factor(c(1,NA))), the first fixed effect parameter is modeled while the second parameter is kept at its initial value. See [TMB::MakeADFun()] for more details.

size

number of trials in binomial model.

dofit

fit the model (`TRUE`) or return the processed data without fitting (`FALSE`)?

silent

silent or include optimization details? Helpful to set to `FALSE` for models that take a while to fit.

Value

An object (list) of class `combTMB`. With a number of useful elements among them:

tmb_obj: The TMB object created by [TMB::MakeADFun()]

tmb_report: output from [TMB::sdreport()]

Gradients: log likelihood gradients with respect to each fixed effect

fit: output from [stats::nlminb()] or [stats::optim()]

data: the fitted data

tmb_params: The parameters list passed to [TMB::MakeADFun()]

...

Details

Binomial families

For a binomial family, the response variable can be in four ways:

(i) the response can be a factor (and the model ranks the first level versus all others);

(ii) the response can be binary (0/1);

(iii) the response can be an array of the form `cbind(success, failure)`; and

(iv) the response can be in proportions, and the 'weights' argument is used to specify the binomial size(N) parameter (`prop ~ ..., weights = N`).

Marginalization

The option of marginalizing the regression parameters is currently only available for the [binomial()], [poisson()] and poigamma() only one random effect.

References

Molenberghs, G., Verbeke, G., and Demetrio, C. G. (2007). An extended random-effects approach to modeling repeated, overdispersed count data. Lifetime data analysis, 13(4), 513-531.

Molenberghs, G., Verbeke, G., Demetrio, C. G., and Vieira, A. M. (2010). A family of generalized linear models for repeated measures with normal and conjugate random effects. Statistical science, 25(3), 325-347.

Molenberghs, G., Verbeke, G., Iddi, S., and Demetrio, C. G. (2012). A combined beta and normal random-effects model for repeated, overdispersed binary and binomial data. Journal of Multivariate Analysis, 111, 94-109.

Iddi, S. and Molenberghs, G. (2012). A combined overdispersed and marginalized multilevel model. Computational Statistics & Data Analysis, 56(6), 1944-1951.

Molenberghs, G., Verbeke, G., and Demetrio, C. G. (2017). Hierarchical models with normal and conjugate random effects: a review. Statistics and Operations Research Transactions, 41(2), 191-253.

Examples

# \donttest{
library(combTMB)
#Poisson-Normal model
m1 <- combTMB(OT ~ Period+(1|Donor), family=poisson(), data=embryos)
summary(m1)
#> Family: poisson 
#> Link function: log 
#> Formula: OT ~ Period + (1 | Donor) 
#> Number of obs: 1148 
#> -2 x logLik         AIC         BIC    df.resid 
#>      7839.5      7845.5      7860.7        1145 
#> 
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  2.69168    0.03307  81.397  < 2e-16 ***
#> PeriodP2     0.06419    0.01859   3.454 0.000553 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Random effects:
#>                   Estimate Std. Error
#> Donor (Intercept)    0.268      0.025
#> 
#> Number of subjects: 318 

#Using 'map' to fix parameter
m1_map <- combTMB(OT ~ Period+(1|Donor), family=poisson(), map=list(beta=factor(c(1,NA))),
                  data=embryos)
summary(m1_map)
#> Family: poisson 
#> Link function: log 
#> Formula: OT ~ Period + (1 | Donor) 
#> Number of obs: 1148 
#> -2 x logLik         AIC         BIC    df.resid 
#>      7851.5      7855.5      7865.6        1146 
#> 
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  2.73290    0.03099   88.18   <2e-16 ***
#> PeriodP2     0.00000         NA      NA       NA    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Random effects:
#>                   Estimate Std. Error
#> Donor (Intercept)   0.2715      0.025
#> 
#> Number of subjects: 318 

#Combined model: Poisson-Gamma-Normal model
m2 <- combTMB(OT ~ Period+(1|Donor), family=poigamma(), data=embryos)
summary(m2)
#> Family: poigamma 
#> Link function: log 
#> Formula: OT ~ Period + (1 | Donor) 
#> Dformula: ~1 
#> Number of obs: 1148 
#> -2 x logLik         AIC         BIC    df.resid 
#>      7650.5      7658.5      7678.7        1144 
#> 
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  2.69904    0.03475  77.662  < 2e-16 ***
#> PeriodP2     0.06813    0.02520   2.704  0.00686 ** 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#>  
#> 
#> Random effects:
#>                   Estimate Std. Error
#> Donor (Intercept)   0.2484      0.024
#> Overdisp.(theta)   21.9242      2.498
#> 
#> Number of subjects: 318 

#Combined overdispersed and marginalized multilevel models (COMMM)
m2_COMMM <- combTMB(OT ~ Period+(1|Donor), family=poigamma(), doMarginal=TRUE,data=embryos)
summary(m2_COMMM)
#> Family: poigamma 
#> Link function: log 
#> Formula: OT ~ Period + (1 | Donor) 
#> Dformula: ~1 
#> Number of obs: 1148 
#> -2 x logLik         AIC         BIC    df.resid 
#>      7650.5      7658.5      7678.7        1144 
#> 
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  2.82325    0.03628  77.828  < 2e-16 ***
#> PeriodP2     0.06813    0.02520   2.704  0.00686 ** 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#>  
#> 
#> Random effects:
#>                   Estimate Std. Error
#> Donor (Intercept)   0.2484      0.024
#> Overdisp.(theta)   21.9243      2.498
#> 
#> Number of subjects: 318 

#Logistic-Normal model
m3 <- combTMB(onyresp~treatn-1+treatn%in%time+(1|idnum), family = binomial(), data = toenail)
summary(m3)
#> Family: binomial 
#> Link function: logit 
#> Formula: onyresp ~ treatn - 1 + treatn %in% time + (1 | idnum) 
#> Number of obs: 1908 
#> -2 x logLik         AIC         BIC    df.resid 
#>      1252.6      1262.6      1290.4        1903 
#> 
#>              Estimate Std. Error z value Pr(>|z|)    
#> treatn0      -2.56253    0.81191  -3.156  0.00160 ** 
#> treatn1      -2.82287    0.86546  -3.262  0.00111 ** 
#> treatn0:time -0.41415    0.04890  -8.470  < 2e-16 ***
#> treatn1:time -0.57672    0.06419  -8.984  < 2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Random effects:
#>                   Estimate Std. Error
#> idnum (Intercept)    21.21      6.853
#> 
#> Number of subjects: 294 

#Combined model: Beta-Bernoulli-Normal model
m4 <- combTMB(onyresp~treatn-1+treatn%in%time+(1|idnum),
      family = betabernoulli(link = "probit"),data = toenail,
       control = combTMBcontrol(optimizer = "nlminb",newtonLoops = 2,
                              lower = c(betad=0.1),upper = c(betad=Inf)))
#> Warning: The model may not have converged. Maximum final gradient: 4.46069636421333.
summary(m4)
#> Family: betabernoulli 
#> Link function: probit 
#> Formula: onyresp ~ treatn - 1 + treatn %in% time + (1 | idnum) 
#> Dformula: ~1 
#> Number of obs: 1908 
#> -2 x logLik         AIC         BIC    df.resid 
#>      1140.7      1152.7      1186.1        1902 
#> 
#>              Estimate Std. Error z value Pr(>|z|)    
#> treatn0       -4.5585     0.2954 -15.430  < 2e-16 ***
#> treatn1       -4.6130     0.3083 -14.961  < 2e-16 ***
#> treatn0:time  -1.7995     0.1513 -11.896  < 2e-16 ***
#> treatn1:time  -2.1381     0.3251  -6.577  4.8e-11 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#>  
#> 
#> Random effects:
#>                   Estimate Std. Error
#> idnum (Intercept) 453.8969    119.170
#> Overdisp.(theta)    0.2148      0.026
#> 
#> Number of subjects: 294 
# }