Optimization control options

Optimization control options

Usage

combTMBcontrol(
  optimizer = c("nlminb", "optim"),
  optim.method = "BFGS",
  eval.max = 2000L,
  iter.max = 2000L,
  maxit = 500L,
  nlminbLoops = 1L,
  newtonLoops = 0L,
  zerodisp_val = log(sqrt(.Machine$double.eps)),
  start_params = NULL,
  lower = NULL,
  upper = NULL,
  gjp = TRUE,
  ...
)

Arguments

optimizer

optimization with "nlminb" or "optim".

optim.method

by default, the optimizer performs minimization with the `BFGS` method. If the lower and upper limits are declared the indicated method is the `L-BFGS-B`.

eval.max

maximum number of evaluations of the objective function allowed.

iter.max

maximum number of iterations allowed.

maxit

the maximum number of iterations for the "optim" method.

nlminbLoops

how many times to run "nlminb" optimization. Sometimes restarting the optimizer at the previous best values aids convergence. If the maximum gradient is still too large, try increasing this to `2`.

newtonLoops

how many Newton optimization steps to try with "optimHess" after running "nlminb". Sometimes aids convergence.

zerodisp_val

value of the dispersion parameter when dformula=~0 is specified.

start_params

the user can specify initial values for the parameters.

lower

an optional named list of lower bounds within the optimization.

upper

an optional named list of upper bounds within the optimization.

gjp

logical. Passed to `getJointPrecision` in "sdreport". Must be `TRUE` to use simulation-based methods. If not needed, setting this `FALSE` will reduce object size.

...

currently not used.

Author

Deoclecio Jardim Amorim <deocleciojardimamorim@gmail.com>.