Adjust glm models

Adjust glm models

Usage

combTMBglm(
  formula,
  family = gaussian(link = "identity"),
  data = NULL,
  starting_val = "zero",
  contrasts = NULL,
  weights = NULL,
  size = NULL,
  offset = NULL,
  dofit = TRUE,
  map = NULL,
  silent = TRUE,
  control = combTMBcontrol(),
  na.action
)

Arguments

formula

An object of class "formula" describe the model for mean.

family

An character indicating the model used. Options are model Poisson:"poisson", negative binomial type 2 "nb2", and binomial "binomial" - see Details.

data

An optional data frame containing the variables in the model. If missing data, the variables ara taken from environment(formula).

starting_val

An character indicating which method should be used to provide initial values. Options are the estimated values through a fitted model using glm.fit "fm" and all initial parameter values are zero "zero".

contrasts

An optional list. See the contrasts.arg of model.matrix.default.

weights

An optional vector of "prior weights" to be used in the fitting process. Should be NULL or a numeric vector - see Details.

size

Number of trials in binomial model.

offset

A numeric vector representing the model offset.

dofit

Fit the model (`TRUE`) or return the processed data without fitting (`FALSE`)?

map

xxx.

silent

Silent or include optimization details? Helpful to set to `FALSE` for models that take a while to fit.

control

Optimization control options via combTMBcontrol.

na.action

a function that specifies how to handle observations containing NAs. The default action (na.omit, inherited from the 'factory fresh' value of getOption("na.action")) strips any observations with any missing values in any variables. Using na.action = na.exclude

Value

retorna a um objeto da classe combTMBglm.

Details

Link function: Options are link="log", link="logit", link="probit" and link="cloglog".

binomial models can be specified typically a two-column matrix cbind(successes,failures)~... form or prob ~ ..., weights = N.

In the case of the family gaussian. Weights \(w_i\) are normalized by multiplying them by \(\frac{n}{\sum_{i=1}^{n}w_i}\), where n is the sample size. Thus, the estimated covariance matrix of the parameter estimators is scale invariant of the `weights` variable.

Author

Deoclecio Jardim Amorim <deocleciojardimamorim@gmail.com>